[ISSUE #22] user biz develop in the second edition of sduoj (#23)
* [ISSUE #22] new db model

* [ISSUE #22] emailVerify API

* authentication in gateway, 'all' role mean URL permission is public.

* Edit the priority of ResponseHandlers

* [ISSUE #22] forgetPassword, resetPassword API.

+++ sduoj-submit/sduoj-submit-service/src/main/java/cn/edu/sdu/qd/oj/submit/service/SubmitService.java
@@ -101,7 +101,7 @@ public class SubmitService {
     }
 
     public PageResult<SubmissionListDTO> querySubmissionByPage(String username, Integer problemId, int pageNow, int pageSize) {
-        Integer userId = null;
+        Long userId = null;
         if (StringUtils.isNotBlank(username)) {
             try {
                 userId = userClient.queryUserId(username);
